Make test harness more robust.
To get around the fact that the hashing function for tape name
generation was taking headers into account, the test harness was using
a singleton hashing function that only let one request be recorded
ever. This replaces the hashing function with the original minus the
portion that considers headers.

Also, all of the test support functions were extracted into their own
files so that the test can stand alone.