Back to using the already-read JSON Schema
Unforrunately, it seems we don't have access to the filename of the
schema that has been included, only its content and the logical name
by which it's known. So when the RAML schema says:

  schemas:
    - instances: !include instances.json

We get given the logical name "instances" and the content of the file
"instances.json" -- but NOT the actual filename "instances.json" which
we need to give to the resolver.

The present version of the code attempts to futz around this by
setting an ID into each schema that specifies the base-path to use for
file resolution. But that's not getting the job done.

Man, what a lot of running around in circles.