Move raml2graphqlType back into gatherAPI.js
The gathered API structure already uses GraphQL-style query-names, and
we want it to use GraphQL-style fundamental type names for consistency
-- also because we will shortly introduce schema-defined types, and
will want the gathered API object to use GraphQL-style type names for
those, too.

One reason for using GraphQL-style names in the gathered API object
instead of RAML/XML Schema-style names is that this way they only need
to be translated in one place (`gatherAPI.js`) rather than two
(`asSchema.js` and `asResolvers.js`). That makes the code shorter, and
eliminates the possibility of accidental inconsistency between the
behaviour of the two `as*.js` files.