Catch and display errors from mod-graphql server
In fact there are no errors when running this test; but when I ran
into errors developing a test of the instances test, they were not
getting displayed. So having updated that (not-yet-committed) test to
emit diagnostics, I back-ported the same change to this test. Down the
line, the code that makes the GraphQL request will be abstracted out
of the tests, into a library.