Distinguish between response-spec absent and schema absent
The intent of the present code, as discussed in MODGQL-41, is to
simply omit resources that include no specification for a response of
type application/json, rather than throwing an error. And error _is_
still thrown if such a specification is included, but it does not
specify the schema of the JSON response.

Needs re-indenting, but I am doing this intermediate commit so the
substantive code-changes are easy to see.