Parse referenced schemas by hand
Completely change the way we handle RAMLs with JSON Schemas that
include other JSON Schemas.

Previously we expanded them all in a single shot using
$RefParser.dereference(), which resulted in duplicate copies of
schemas that were referenced more than once, and completely precluded
mutual-reference loops. (Also, relative paths didn't work right.)

Now, instead, we walk through each of the JSON Schemas that is
directly included by the RAML, and recursively parse any further JSON
Schemas that it references. Each such schema is separately registered
so that subsequent references to it can be immediately resolved
without a redundant re-parse -- and so that reference loops work.

This is now mostly working, but there are a couple of wrinkles to be
handled before we're done:
* Test 16 fails with an ENOENT exception, most likely because we're
  not yet handling relative reference paths.
* Test 19 is no longer translating a hyphen in the type name into an
  underscore for some reason.
* Some other tests are yielding changed output because the new code
  names included schemas differently. I need to manually check these
  for equivalence before updating regression-test expectations.

Towards MODGQL-64.