Nail down specific SHA1s for tested RAMLs
This requires known versions not only of mod-inventory-storage but
also the included raml-utils.

At some point, we will want to upgrade mod-graphql so it can correctly
deal with the way these modules use RAML 1.0 (which is not the same as
the way as how the eHoldings module's RAML 1.0 worked). But that's
another problem for a different day.

Meanwhile, this should fix MODGQL-69.