Major change to top-level API
app.js now exports a modGraphql function, which starts the service
when invoked; previously, app.js would directly start the service and
return a handle on the running app.

The main purpose of this change is so that parameters can be passed
into the modGraphql invocation -- specifically, the paths to the RAML
files which are to be converted.

Fixes MODGQL-79.