Depend on EXACTLY version 1.1.47 of raml-1-parser
Previously we accepted anything ^1.1.42, but as of 1.1.48 the
schema-translation tests fail because the RAML parser is now using a
stricter interpretation of JSON Schema in which all boolean fields
must be required. I can't tell exactly what change they made to cause
this, as there seems to be no changelog. Further investigations will
follow, but for now this change means tests pass again.