Remove alternativeTitles from test
It looks like its structure has changed since the test was written --
perhaps it now an array of objects rather than an array of
strings. Testing for it was never the point of this test anyway, so
I've just removed it so that the test should work with both old and
new versions of the back-end module.