create-schemas supports optional module-specific `ramlPath`
When this is specified in one of the module stanzas of the
`schemaconf.json` file, the downloaded copy's `ramls` directory is a
copy not of the original's `ramls`, but of the specified directory. As
well and being generally more flexible, this is a step towards
supporting modules whose APIs are expressed in OpenAPI and follow the
corresponding conventions.

Part of MODGQL-153.