create-schemas supports optional module-specific `copyFiles`
When this is specified in one of the module stanzas of the
`schemaconf.json` file, each file specified in the array value is
copied into the downloaded copy's `ramls` directory.  This gives us a
way to provide RAML files for modules whose APIs are expressed in
OpenAPI.

Part of MODGQL-153.