Bump interface version to 1.3
This may not be in some sense technically correct, since the change
here is not exactly to what response is returned for a given
request. But it's the right thing pragmatically, as client code needs
to be able to express a dependency on the new way of interpreting
certain JSON schemas.