Separate test and build directories (#93)
A new, self-contained [`build` directory](build) is used for building the Docker container. This is now distinct from the [`tests`](tests) directory used for running tests (as it always should have been) and from the schema-configuration file and hand-crafted `mod-search` RAML that are used in development of the [`create-schemas`](create-schemas) part of the software.

Fixes MODGQL-158.

Note that at this point, the Dockerfile created from the `build` directory does not run successfully — see MODGQL-165 — but this is essentially a separate issue. Having fixed the present issue will make one clearer to think about.