Revert "Move all babel 6 deps to devDependencies (#115)"
This reverts commit eb2821b71e9f9a4b01dcc3b905acd122e7c6636b.

The `yarn start` command (which is invoked as the Docker container's
`RUN` command) uses babel, so that translation happens at startup
time. So that obviously fails now Babel is just a dev-dependency.

There are various better approaches we could take, but for now the
main thing is just to make it work again,